feat(core): add NewEmptyPageInfo and refactor file organization
Added NewEmptyPageInfo() constructor to provide a convenient way to
create empty PageInfo instances for error cases and empty result sets.

Also refactored by splitting models.go into page_args.go and page_info.go
to improve code organization and make components easier to locate.
